feat: file structure update and cleanup
